Moved some directories into app directory
