+++ src/app/components/wrapper/BgWrapper.tsx
@@ -1,5 +1,5 @@
 import { Box } from "@mui/material";
-import ImageUrl from "@/public/images/top-bg-image.jpg";
+import ImageUrl from "@/src/app/public/images/top-bg-image.jpg";
 
 export const BgWrapper: React.FC<{ children: React.ReactNode }> = ({
   children,
